requirements.txt
PyQt5==5.15.6
python-mpv==1.0.1
